+++ header.php
@@ -3,15 +3,12 @@
 <head>
 <meta http-equiv="content-type" content="text/html; charset=<?php bloginfo('charset'); ?>">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
-<meta name="description" content="">
-<meta name="keywords" content="">
-<meta name="author" content="">
 <?php wp_head(); ?>
 </head>
 <?php $awada_theme_options = awada_theme_options();
 $class = "";
 if ($awada_theme_options['site_layout'] == 'boxed') {
-    $class = 'boxed';
+    $class .= ' boxed ';
 } ?>
 <body <?php body_class($class); ?>>
 <?php if ($awada_theme_options['site_layout'] == 'boxed') { ?>
@@ -24,19 +21,19 @@ if ($awada_theme_options['site_layout'] == 'boxed') {
 		<div class="col-lg-4 col-md-4 col-sm-4 col-xs-12" id="social-icons">
 			<div class="social-icons">
 				<?php if ($awada_theme_options['social_facebook_link'] != '') { ?>
-				<span><a data-toggle="tooltip" data-placement="bottom" title="Facebook" href="<?php echo esc_url($awada_theme_options['social_facebook_link']); ?>" target="_blank"><i class="fa fa-facebook"></i></a></span>
+				<span><a id="facebook" data-toggle="tooltip" data-placement="bottom" title="Facebook" href="<?php echo esc_url($awada_theme_options['social_facebook_link']); ?>" target="_blank"><i class="fa fa-facebook"></i></a></span>
 				<?php } if ($awada_theme_options['social_google_plus_link'] != '') { ?>
-				<span><a data-toggle="tooltip" data-placement="bottom" title="Google Plus" href="<?php echo esc_url($awada_theme_options['social_google_plus_link']); ?>" target="_blank"><i class="fa fa-google-plus"></i></a></span>
+				<span><a id="googleplus" data-toggle="tooltip" data-placement="bottom" title="Google Plus" href="<?php echo esc_url($awada_theme_options['social_google_plus_link']); ?>" target="_blank"><i class="fa fa-google-plus"></i></a></span>
 				<?php } if ($awada_theme_options['social_twitter_link'] != '') { ?>
-				<span><a data-toggle="tooltip" data-placement="bottom" title="Twitter" href="<?php echo esc_url($awada_theme_options['social_twitter_link']); ?>" target="_blank"><i class="fa fa-twitter"></i></a></span>
+				<span><a id="twitter" data-toggle="tooltip" data-placement="bottom" title="Twitter" href="<?php echo esc_url($awada_theme_options['social_twitter_link']); ?>" target="_blank"><i class="fa fa-twitter"></i></a></span>
 				<?php } if ($awada_theme_options['social_youtube_link'] != '') { ?>
-				<span><a data-toggle="tooltip" data-placement="bottom" title="Youtube" href="<?php echo esc_url($awada_theme_options['social_youtube_link']); ?>" target="_blank"><i class="fa fa-youtube"></i></a></span>
+				<span><a id="youtube" data-toggle="tooltip" data-placement="bottom" title="Youtube" href="<?php echo esc_url($awada_theme_options['social_youtube_link']); ?>" target="_blank"><i class="fa fa-youtube"></i></a></span>
 				<?php } if ($awada_theme_options['social_linkedin_link'] != '') { ?>
-				<span><a data-toggle="tooltip" data-placement="bottom" title="Linkedin" href="<?php echo esc_url($awada_theme_options['social_linkedin_link']); ?>" target="_blank"><i class="fa fa-linkedin"></i></a></span>
+				<span><a id="linkedin" data-toggle="tooltip" data-placement="bottom" title="Linkedin" href="<?php echo esc_url($awada_theme_options['social_linkedin_link']); ?>" target="_blank"><i class="fa fa-linkedin"></i></a></span>
 				<?php } if ($awada_theme_options['social_dribbble_link'] != '') { ?>
-				<span><a data-toggle="tooltip" data-placement="bottom" title="Dribbble" href="<?php echo esc_url($awada_theme_options['social_dribbble_link']); ?>" target="_blank"><i class="fa fa-dribbble"></i></a></span>
+				<span><a id="dribbble" data-toggle="tooltip" data-placement="bottom" title="Dribbble" href="<?php echo esc_url($awada_theme_options['social_dribbble_link']); ?>" target="_blank"><i class="fa fa-dribbble"></i></a></span>
 				<?php } if ($awada_theme_options['social_skype_link'] != '') { ?>
-				<span class="last"><a data-toggle="tooltip" data-placement="bottom" title="Skype" href="<?php echo esc_url($awada_theme_options['social_skype_link']); ?>" target="_blank"><i class="fa fa-skype"></i></a></span>
+				<span class="last"><a id="skype" data-toggle="tooltip" data-placement="bottom" title="Skype" href="<?php echo esc_url($awada_theme_options['social_skype_link']); ?>" target="_blank"><i class="fa fa-skype"></i></a></span>
 				<?php } ?>
 			</div><!-- end social icons -->
 		</div><!-- end columns -->
@@ -56,7 +53,7 @@ if ($awada_theme_options['site_layout'] == 'boxed') {
 </div><!-- end topbar -->
 <?php } ?>
 <header id="header-style-1" style="<?php if (get_header_image()) : ?> background-image:url('<?php header_image();?>'); <?php endif; ?>" class="<?php echo esc_attr($awada_theme_options['menubarcolor']); ?> <?php if($awada_theme_options['headersticky']=='0'){ echo 'affix1'; } ?>">
-	<div class="container">
+	<div id="header_menu" class="container">
 		<nav class="navbar yamm navbar-default">
 			<div class="navbar-header">
 				<button type="button" data-toggle="collapse" data-target="#navbar-collapse-1" class="navbar-toggle">
@@ -64,16 +61,18 @@ if ($awada_theme_options['site_layout'] == 'boxed') {
 					<span class="icon-bar"></span>
 					<span class="icon-bar"></span>
 				</button>
-					<a href="<?php echo esc_url(home_url('/')); ?>" class="navbar-brand"><?php
-					if (get_custom_logo()!='') {
+				<?php
+					if (function_exists(the_custom_logo())) {
 						the_custom_logo();
-					} else { ?>
-						<p><?php echo get_bloginfo('name'); ?> </p><?php
+					}?>
+					<a href="<?php echo esc_url(home_url('/')); ?>" class="navbar-brand"><?php
+					if (display_header_text()) {?>
+						<p id="logo_text_id"><?php echo get_bloginfo('name'); ?> </p><?php
 					} ?>
 					</a>	
 			</div><!-- end navbar-header -->
 			
-			<div id="navbar-collapse-1" class="navbar-collapse collapse navbar-right <?php if ($awada_theme_options['logo_layout']=='right'){ echo 'style="float:right;"'; } //if ($awada_theme_options['header_layout']==1) { echo 'navbar-right'; } ?>">
+			<div id="navbar-collapse-1" class="navbar-collapse collapse navbar-right <?php if ($awada_theme_options['logo_layout']=='right'){ echo 'style="float:right;"'; } ?>">
 					<?php wp_nav_menu(array(
 								'theme_location' => 'primary',
 								'menu_class' => 'nav navbar-nav',