+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+###Version 0.3
+* removed unused files.
+* Recommended kirki plugin.  
+* Minor bug fixed.
 ###Version 0.2
 * Text domain issue fixed.
 * All other issue fixed.